requirements.txt
# requirements.txt

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
pymongo==4.6.0
motor==3.3.2

# Vector Database
chromadb==0.4.18

# LangChain
langchain==0.0.350
langchain-community==0.0.3
langchain-core==0.1.0

# Embeddings & Models
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2

# API Clients
groq==0.4.1
requests==2.31.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Text Processing
nltk==3.8.1
spacy==3.7.2

# Audio Processing
pydub==0.25.1

# Translation
googletrans==4.0.0rc1

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Web Interface
streamlit==1.29.0
streamlit-chat==0.1.1
streamlit-mic-recorder==0.0.4
# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
httpx==0.26.0
aiofiles==23.2.1

# Cross-Encoder for Reranking
sentence-transformers-cross-encoder==0.1.0